housekeeping: moved recipe component into ViewRecipe page
