+++ src/meal-ui/src/pages/ViewRecipe/RecipeCard/index.js
